# openapi.yaml
openapi: 3.0.0
info:
  version: 0.0.0
  title: Conquery API
  description: Conquery API
  license:
    #TODO find a fitting license
    name: TODO
    url: http://TODO/
servers:
  - url: "http://localhost:8081"
...
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic

    Bearer:
      type: http
      scheme: bearer
  parameters:
    Dataset:
      in: path
      name: dataset
      required: true
      schema:
        $ref: "#/components/schemas/DatasetId"
...
  schemas:
    Id:
      description: Typed identifier of an object. Usually consists of a recursive structure separated by dots, starting with the dataset. The dataset can be omitted, when the url is already in a context with a bound dataset.
      type: string
...
    DatasetId:
      $ref: "#/components/schemas/Id"
    FilterId:
      $ref: "#/components/schemas/Id"
...
    FilterValueBase:
      type: object
      properties:
        filter:
          $ref: "#/components/schemas/FilterId"
    FilterValueMultiSelect:
      allOf:
        - $ref: "#/components/schemas/FilterValueBase"
      properties:
        type:
          type: string
          enum:
            - "MULTI_SELECT"
        value:
          type: array
          items:
            type: string
    FilterValueBigMultiSelect:
      allOf:
        - $ref: "#/components/schemas/FilterValueBase"
      properties:
        type:
          type: string
          enum:
            - "BIG_MULTI_SELECT"
        value:
          type: array
          items:
            type: string
    FilterValueSingleSelect:
      allOf:
        - $ref: "#/components/schemas/FilterValueBase"
      properties:
        type:
          type: string
          enum:
            - "SELECT"
        value:
          type: string
    FilterValueString:
      allOf:
        - $ref: "#/components/schemas/FilterValueBase"
      properties:
        type:
          type: string
          enum:
            - "STRING"
        value:
          type: string
    FilterValueInteger:
      allOf:
        - $ref: "#/components/schemas/FilterValueBase"
      properties:
        type:
          type: string
          enum:
            - "INTEGER"
        value:
          type: integer
    FilterValueNumberRange:
      allOf:
        - $ref: "#/components/schemas/FilterValueBase"
      properties:
        type:
          type: string
          enum:
            - "INTEGER_RANGE"
            - "DECIMAL_RANGE"
            - "MONEY_RANGE"
            - "REAL_RANGE"
        value:
          type: object
          properties:
            min:
              type: number
            max:
              type: number
    FilterValueIntegerRange:
      allOf:
        - $ref: "#/components/schemas/FilterValueBase"
      properties:
        type:
          type: string
          enum:
            - "INTEGER_RANGE"
        value:
          type: object
          properties:
            min:
              type: integer
            max:
              type: integer
...
    FilterValue:
      description: "Reference to a filter, and parameters for filtering with it.
      Note that the filter is included by reference, and its actual logic is described in the defining concept."
      type: object
      discriminator:
        propertyName: "type"
        mapping:
          "INTEGER_RANGE": "#/components/schemas/FilterValueIntegerRange"
          "REAL_RANGE": "#/components/schemas/FilterValueNumberRange"
          "MONEY_RANGE": "#/components/schemas/FilterValueNumberRange"
          "DECIMAL_RANGE": "#/components/schemas/FilterValueNumberRange"
          "INTEGER": "#/components/schemas/FilterValueInteger"
          "STRING": "#/components/schemas/FilterValueString"
          "SELECT": "#/components/schemas/FilterValueSingleSelect"
          "BIG_MULTI_SELECT": "#/components/schemas/FilterValueMultiSelect"
          "MULTI_SELECT": "#/components/schemas/FilterValueBigMultiSelect"
      oneOf:
        - $ref: "#/components/schemas/FilterValueNumberRange"
        - $ref: "#/components/schemas/FilterValueNumberRange"
        - $ref: "#/components/schemas/FilterValueNumberRange"
        - $ref: "#/components/schemas/FilterValueNumberRange"
        - $ref: "#/components/schemas/FilterValueInteger"
        - $ref: "#/components/schemas/FilterValueString"
        - $ref: "#/components/schemas/FilterValueSingleSelect"
        - $ref: "#/components/schemas/FilterValueMultiSelect"
        - $ref: "#/components/schemas/FilterValueBigMultiSelect"
...
paths:
...
  /datasets/{dataset}/queries/resolve-entities:
    parameters:
      - $ref: "#/components/parameters/Dataset"
    post:
      tags:
        - entity-preview
      operationId: resolve-entities
      summary: Search for a few entities, returning their ids.
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/FilterValue"
      responses:
        "200":
          description: List of tuple of ids.
          content:
            application/json:
              example: [ { "kind1": "id1", "kind2": "id2" }, { "kind1": "id3", "kind2": "id4" } ]
              schema:
                type: object
                additionalProperties:
                  type: string
...